fix(http-client-csharp): support int64 enum/union base types
Extend InputEnumTypeValueConverter to handle InputPrimitiveTypeKind.Int64
by reading values via GetInt64() and widening InputEnumTypeIntegerValue's
IntegerValue from int to long. Also include the offending kind, enum,
and value names in the default-arm JsonException message so unsupported
kinds are diagnosable from build output.

Fixes Azure/azure-sdk-for-net#59168.

+++ packages/http-client-csharp/generator/Microsoft.TypeSpec.Generator.Input/src/InputTypes/InputEnumTypeIntegerValue.cs
@@ -5,12 +5,12 @@ namespace Microsoft.TypeSpec.Generator.Input
 {
     internal class InputEnumTypeIntegerValue : InputEnumTypeValue
     {
-        public InputEnumTypeIntegerValue(string name, int integerValue, InputPrimitiveType valueType, string? summary, string? doc, InputEnumType? enumType = default)
+        public InputEnumTypeIntegerValue(string name, long integerValue, InputPrimitiveType valueType, string? summary, string? doc, InputEnumType? enumType = default)
             : base(name, integerValue, valueType, summary, doc, enumType)
         {
             IntegerValue = integerValue;
         }
 
-        public int IntegerValue { get; }
+        public long IntegerValue { get; }
     }
 }
